refactor(SPV-1278): move test helpers from steps file into regression test file.
